Add files via upload
This file contains 3 dashboards . The Dashboards are created in such a way that is easy to read for the user. 2 of the dashboards namely customer service and finance dashboards are made in dark theme and the order fulfillment dashboard is created with a light theme in the background to showcase my dynamic skills in creating dashboards. Really looking forward to hear further suggestions to increase my skills for the better use.